pyproject.toml
[project]
name = "places-project"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "requests",
  "python-dotenv",
  "backoff",
  "SQLAlchemy>=2.0",
  "pydantic>=2.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
lint.select = [
  "E", "F", "I", "UP", "B", "SIM", "C4", "RET", "RUF"
]
lint.ignore = ["ANN101", "ANN102"]
exclude = [".venv", ".mypy_cache", ".ruff_cache", "build", "dist", "__pycache__"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
strict_optional = true
disallow_untyped_defs = true
ignore_missing_imports = true
